Load InboxViewController based on propper inbox path
